attendanceAgreeControllerとattendanceAgree.htmlの変更

+++ kintai_app/src/main/java/com/kintai/Controller/AttendanceAgreeController.java
@@ -2,6 +2,8 @@ package com.kintai.Controller;
 
 import java.util.List;
 
+import jakarta.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.kintai.Dao.StampsDAO;
 import com.kintai.Entity.StampsEntity;
 import com.kintai.Service.AttendanceAgreeService;
+import com.kintai.Service.AuthService;
 
 @Controller
 public class AttendanceAgreeController {
@@ -21,10 +24,18 @@ public class AttendanceAgreeController {
 	private AttendanceAgreeService attendanceAgreeService;
 	@Autowired
 	private StampsDAO stampsDAO;
+	@Autowired
+    private AuthService authService;
 
 	//勤怠変更申請ページ遷移＆一覧
 	@GetMapping("/attendanceAgree")
-	public String attendanceAgree(Model model) {
+	public String attendanceAgree(HttpSession session, Model model) {
+        
+        //  管理者Session権限验证
+    	
+		 if (!authService.isManagerLoggedIn(session)) {
+	            return "redirect:/auth";
+	        }
 		List<StampsEntity> changeRequests = attendanceAgreeService.getAttendanceAgreeList();
 		model.addAttribute("changeRequests", changeRequests);
 		return "attendanceAgree";
@@ -33,9 +44,26 @@ public class AttendanceAgreeController {
 	//打刻変更の承認ajax
 	@PostMapping("/approve-request")
 	@ResponseBody
-	public String approveRequest(@RequestParam Long stampId) {
-
+	public String approveRequest(@RequestParam Long stampId, HttpSession session) {
+		
+		 //  管理者Session権限验证
+		if (!authService.isManagerLoggedIn(session)) {
+            return "redirect:/auth";
+        }
 		stampsDAO.deleteAttendanceAgree(stampId);
+		return "success";
+	}
+	
+	//打刻変更の却下ajax
+	@PostMapping("/deny-request")
+	@ResponseBody
+	public String denyRequest(@RequestParam Long stampID, HttpSession session) {
+		
+		 //  管理者Session権限验证
+		if (!authService.isManagerLoggedIn(session)) {
+            return "redirect:/auth";
+        }
+		stampsDAO.deleteAttendanceAgree(stampID);
 		return "success";
 	}
